Change solve to return long

+++ src/Problem.java
@@ -2,8 +2,8 @@ public abstract class Problem {
 
     /**
      * Solves this problem from ProjectEuler.
-     * @return the integer solution to this problem
+     * @return the solution to this problem
      */
-    public abstract int solve();
+    public abstract long solve();
 
 }
